refactor wallet module
