+++ api/server/server.go
@@ -50,7 +50,7 @@ import (
 	reputationRpc "github.com/textileio/powergate/reputation/rpc"
 	txndstr "github.com/textileio/powergate/txndstransform"
 	"github.com/textileio/powergate/util"
-	"github.com/textileio/powergate/wallet"
+	walletModule "github.com/textileio/powergate/wallet/module"
 	walletRpc "github.com/textileio/powergate/wallet/rpc"
 	"google.golang.org/grpc"
 )
@@ -73,7 +73,7 @@ type Server struct {
 	mi *minerModule.Index
 	fi *faultsModule.Index
 	dm *dealsModule.Module
-	wm *wallet.Module
+	wm *walletModule.Module
 	rm *reputation.Module
 	nm pgnet.Module
 	hm *health.Module
@@ -174,7 +174,7 @@ func NewServer(conf Config) (*Server, error) {
 	if err != nil {
 		return nil, fmt.Errorf("creating deal module: %s", err)
 	}
-	wm, err := wallet.New(c, masterAddr, conf.WalletInitialFunds, conf.AutocreateMasterAddr)
+	wm, err := walletModule.New(c, masterAddr, conf.WalletInitialFunds, conf.AutocreateMasterAddr)
 	if err != nil {
 		return nil, fmt.Errorf("creating wallet module: %s", err)
 	}
