+++ ffs/integrationtest/integration_test.go
@@ -35,7 +35,7 @@ import (
 	"github.com/textileio/powergate/tests"
 	txndstr "github.com/textileio/powergate/txndstransform"
 	"github.com/textileio/powergate/util"
-	"github.com/textileio/powergate/wallet"
+	walletModule "github.com/textileio/powergate/wallet/module"
 )
 
 const (
@@ -1383,7 +1383,7 @@ func newFFSManager(t *testing.T, ds datastore.TxnDatastore, lotusClient *apistru
 	sched, err := scheduler.New(txndstr.Wrap(ds, "ffs/scheduler"), l, hl, cl)
 	require.NoError(t, err)
 
-	wm, err := wallet.New(lotusClient, masterAddr, *big.NewInt(iWalletBal), false)
+	wm, err := walletModule.New(lotusClient, masterAddr, *big.NewInt(iWalletBal), false)
 	require.Nil(t, err)
 
 	pm := paych.New(lotusClient)
