+++ ffs/manager/manager_test.go
@@ -11,7 +11,6 @@ import (
 	"github.com/stretchr/testify/require"
 	paych "github.com/textileio/powergate/paych/lotus"
 	"github.com/textileio/powergate/tests"
-	"github.com/textileio/powergate/wallet"
 )
 
 func TestMain(m *testing.M) {
@@ -128,7 +127,7 @@ func TestDefaultConfig(t *testing.T) {
 
 func newManager(t *testing.T, ds datastore.TxnDatastore) (*Manager, func()) {
 	client, addr, _ := tests.CreateLocalDevnet(t, 1)
-	wm, err := wallet.New(client, addr, *big.NewInt(4000000000), false)
+	wm, err := walletModule.New(client, addr, *big.NewInt(4000000000), false)
 	require.Nil(t, err)
 	pm := paych.New(client)
 	m, err := New(ds, wm, pm, nil)
