+++ wallet/module/wallet.go
@@ -1,4 +1,4 @@
-package wallet
+package module
 
 import (
 	"context"
